collector/requirements.txt
psycopg2-binary==2.9.3
requests==2.28.1
git+https://github.com/wiedehopf/FlightRadar24-API.git@master
